ENH: Add drop_absorbed to AbsorbingLS
Add the ability to drop absorbed to AbsorbingLS

closes #276

+++ linearmodels/shared/io.py
@@ -11,7 +11,7 @@ def _str(v: float) -> str:
     av = abs(v)
     digits = 0
     if av != 0:
-        digits = np.ceil(np.log10(av))
+        digits = int(np.ceil(np.log10(av)))
     if digits > 4 or digits <= -4:
         return "{0:8.4g}".format(v)
